pyproject.toml
[project]
name = "football-data-pipeline"
version = "0.1.0"
description = "Pipeline de datos multi-fuente para fútbol: API-Football + Understat + FBref"
license = { file = "LICENSE" }
authors = [
    { name = "Juanje Marquez", email = "juanje.mn@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "httpx>=0.27.0",
    "soccerdata>=1.8.8",
    "rapidfuzz>=3.0.0",
    "pandas>=2.1.0,<3.0.0",
    "pyarrow>=23.0.1",
    "sqlalchemy>=2.0.48",
    "psycopg2-binary>=2.9.11",
    "datasette>=0.65.2",
    "apache-airflow>=3.1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.15.6",
]

[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
